+++ addon/components/paper-slider.js
@@ -30,28 +30,6 @@ export default Component.extend(FocusableMixin, ColorMixin, {
   step: 1,
   tabindex: 0,
 
-  didInsertElement() {
-    this._super(...arguments);
-
-    this._setupSlider();
-  },
-
-  _setupSlider() {
-    let thumbContainer = this.$('.md-thumb-container').get(0);
-    let sliderHammer = new Hammer(thumbContainer);
-    this._thumbContainerHammer = sliderHammer;
-
-    // Enable dragging the slider
-    sliderHammer.get('pan').set({ threshold: 1 });
-    sliderHammer.on('panstart', run.bind(this, this._dragStart))
-      .on('panmove', run.bind(this, this._drag))
-      .on('panend', run.bind(this, this._dragEnd));
-  },
-
-  trackContainer: computed(function() {
-    return this.$('.md-track-container');
-  }),
-
   activeTrackStyle: computed('percent', function() {
     let percent = this.get('percent') || 0;
     return htmlSafe(`width: ${percent * 100}%`);
@@ -73,8 +51,55 @@ export default Component.extend(FocusableMixin, ColorMixin, {
     return (this.get('value') - min) / (max - min);
   }),
 
+  didInsertElement() {
+    this._super(...arguments);
+    if (!this.get('disabled')) {
+      this._setupHammer();
+    }
+  },
+
+  didUpdateAttrs() {
+    this._super(...arguments);
+
+    if (!this.get('disabled') && !this._hammer) {
+      // if it is enabled and we didn't init hammer yet
+      this._setupHammer();
+    } else if (this.get('disabled') && this._hammer) {
+      // if it is disabled and we did init hammer already
+      this._teardownHammer();
+    }
+  },
+
+  willDestroyElement() {
+    this._super(...arguments);
+    if (this._hammer) {
+      this._teardownHammer();
+    }
+  },
+
+  _setupHammer() {
+    let thumbContainer = this.$('.md-thumb-container').get(0);
+
+    // Enable dragging the slider
+    let containerManager = new Hammer.Manager(thumbContainer);
+    let pan = new Hammer.Pan({ direction: Hammer.DIRECTION_HORIZONTAL, threshold: 10 });
+    containerManager.add(pan);
+
+    containerManager.on('panstart', run.bind(this, this.dragStart))
+      .on('panmove', run.bind(this, this.drag))
+      .on('panend', run.bind(this, this.dragEnd));
+
+    this._hammer = containerManager;
+  },
+
+  _teardownHammer() {
+    this._hammer.destroy();
+    delete this._hammer;
+  },
+
   positionToPercent(x) {
-    return Math.max(0, Math.min(1, (x - this.get('sliderDimensions.left')) / this.get('sliderDimensions.width')));
+    let { left, width } = this.sliderDimensions();
+    return Math.max(0, Math.min(1, (x - left) / width));
   },
 
   percentToValue(x) {
@@ -98,19 +123,31 @@ export default Component.extend(FocusableMixin, ColorMixin, {
   dragging: false,
   enabled: computed.not('disabled'),
 
-  sliderDimensions: computed(function() {
-    return this.get('trackContainer')[0].getBoundingClientRect();
-  }),
+  sliderDimensions() {
+    return this.$('.md-track-container').get(0).getBoundingClientRect();
+  },
+
+  // fix to remove content selection highlight on safari
+  mouseDown(event) {
+    event.preventDefault();
+  },
+
+  click(event) {
+    if (this.get('disabled')) {
+      return;
+    }
+
+    this.setValueFromEvent(event);
+  },
 
   setValueFromEvent(event) {
-    // let exactVal = this.percentToValue(this.positionToPercent(event.deltaX || event.clientX));
     let exactVal = this.percentToValue(this.positionToPercent(event.clientX || event.srcEvent.clientX));
     let closestVal = this.minMaxValidator(this.stepValidator(exactVal));
 
-    this.set('value', closestVal);
+    this.sendAction('onChange', closestVal);
   },
 
-  _dragStart(event) {
+  dragStart(event) {
     if (this.get('disabled')) {
       return;
     }
@@ -119,28 +156,26 @@ export default Component.extend(FocusableMixin, ColorMixin, {
     this.set('dragging', true);
     this.$().focus();
 
-    this.get('sliderDimensions');
-
     this.setValueFromEvent(event);
   },
 
-  _dragEnd() {
-    if (this.get('disabled')) {
-      return;
-    }
-
-    this.beginPropertyChanges();
-    this.set('active', false);
-    this.set('dragging', false);
-    this.endPropertyChanges();
-  },
-
-  _drag(event) {
+  drag(event) {
     if (this.get('disabled') || !this.get('dragging')) {
       return;
     }
 
     this.setValueFromEvent(event);
+  },
+
+  dragEnd() {
+    if (this.get('disabled')) {
+      return;
+    }
+
+    this.setProperties({
+      active: false,
+      dragging: false
+    });
   },
 
   keyDown(event) {
@@ -163,7 +198,7 @@ export default Component.extend(FocusableMixin, ColorMixin, {
 
       newValue = this.get('value') + changeAmount;
 
-      this.set('value', this.minMaxValidator(newValue));
+      this.sendAction('onChange', this.minMaxValidator(newValue));
 
       event.preventDefault();
       event.stopPropagation();